requirements.txt
discord.py==2.1.0
async-timeout==3.0.1
dnspython==1.16.0
PyNaCl==1.3.0
discord-py-slash-command==3.0.3